requirements.txt
python-telegram-bot==21.4
openpyxl==3.1.5
tzdata==2025.1
